moved Ordering, Be, Le to halfling::ordering
